Use selector for ticker matches
