+++ react/gameday2/selectors/index.js
@@ -57,3 +57,54 @@ export const getChatsInDisplayOrder = createSelector(
     return displayOrderChats.sort((a, b) => a.name.localeCompare(b.name))
   }
 )
+
+export const getFireduxData = (state, props) => state.firedux.data
+
+export const getEventKey = (state, props) => props.webcast.key
+
+export const getTickerMatches = createSelector(
+  [getFireduxData, getEventKey],
+  (fireduxData, eventKey) => {
+    var compLevelsPlayOrder = {
+      'qm': 1,
+      'ef': 2,
+      'qf': 3,
+      'sf': 4,
+      'f': 5,
+    }
+    function calculateOrder(match) {
+      return compLevelsPlayOrder[match.comp_level] * 1000000 + match.match_number * 1000 + match.set_number
+    }
+
+    var matches = []
+    if (fireduxData &&
+        fireduxData.events &&
+        fireduxData.events[eventKey] &&
+        fireduxData.events[eventKey].matches) {
+      matches = Object.values(fireduxData.events[eventKey].matches)
+    }
+    matches.sort(function(match1, match2) {
+      return calculateOrder(match1) - calculateOrder(match2);
+    });
+
+    var lastMatch = null
+    var selectedMatches = []
+    for (var i in matches) {
+      var match = matches[i]
+
+      if (match.alliances.red.score == -1 || match.alliances.blue.score == -1) {
+        selectedMatches.push(match);
+      } else {
+        lastMatch = match
+        selectedMatches = []  // Reset selectedMatches if matches get skipped
+      }
+    }
+
+    // Prepend lastMatch to matches
+    if (lastMatch != null) {
+      selectedMatches.unshift(lastMatch)
+    }
+
+    return selectedMatches
+  }
+)
